Clear the actions for the children of a tree view
NSAccessibility seems to set the tree view's actions on its children, but
because actions have no parameters, there's no way to know which cell the action
was performed on, so while the action would trigger the handlers would have to
act on the currently selected item, which is a bit odd if you perform the action
on one item, only for it to affect a different one.

To work around this, the accessibility objects for the children cells return
nil as their actions. If something does need to return a valid action it can
override this accessibilityActionNames method and do the necessary