Correctly handle removing widgets
If we remove widgets from the tree that are ignored by the accessibility system
then we need to do a bit more work to remove the accessibility children.

We need to climb the widget tree gathering all the accessible children that
have been placed into the removed widget's accessibility parent and then remove
those.