Make libAtkCocoa.dylib point to the Current libraries instead of hardcoding them
When linking the library Xcode expands the Current symlink and hardcodes that path for any linked libraries. This
means that when MonoDevelop is running on an older version of Mono it will not work correctly. Fix it by fiddling with
the linked library paths after the libAtkCocoa has been built using install_name_tool

Fixes BXC #58667