Rebuild the column map when the model changes
When the model changes, rebuild the column map. This means we handle the situation where columns are being added before the model is set on the view.
Add all columns to the map and skip any that don't have renderers when generating the column arrays. This lets us add and remove renderers after the column
has already been added to the view